resource-timing/iframe-refresh.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>Resource Timing iframe refresh</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    let t = async_test("Test that iframe refreshes are not observable by the parent");
    function verify_iframe_refresh_not_observable() {
        let entries = performance.getEntriesByType("resource");
        let found_first_iframe = false;
        for (entry of entries) {
            if (entry.name == "{{location[scheme]}}://{{host]}}:{{ports[http][0]}}/resource-timing/resources/iframe-that-refreshes.html") {
                found_first_iframe = true;
            }
            assert_not_equals(entry.name, "{{location[scheme]}}://{{host]}}:{{ports[http][0]}}/resource-timing/resources/refreshed_iframe.html", "Refreshed iframe should not be observable");
        }
        assert_true(found_first_iframe, "Initial iframe should be observable");
        t.done();
    }
    window.addEventListener("message", t.step_func(e=>{
        if (e.data == "refreshed") {
            verify_iframe_refresh_not_observable();
        }
    }));
</script>
<iframe src="{{location[scheme]}}://{{host]}}:{{ports[http][0]}}/resource-timing/resources/iframe-that-refreshes.html">
